+++ src/modules/utils/currentcontrol/mcp4451.h
@@ -22,9 +22,9 @@ class MCP4451 : public DigipotBase {
             delete this->i2c;
         }
 
-        void set_current( int channel, double current )
+        void set_current( int channel, float current )
         {
-            current = min( max( current, 0.0 ), this->max_current );
+            current = min( (float) max( current, 0.0f ), this->max_current );
             currents[channel] = current;
             char addr = 0x58;
             while(channel > 3){
@@ -41,7 +41,7 @@ class MCP4451 : public DigipotBase {
             this->i2c_send( addr, addresses[channel], this->current_to_wiper(current) );
         }
 
-        double get_current(int channel)
+        float get_current(int channel)
         {
             return currents[channel];
         }
@@ -56,12 +56,12 @@ class MCP4451 : public DigipotBase {
             this->i2c->stop();
         }
 
-        char current_to_wiper( double current ){
-            return char(ceil(double((this->factor*current))));
+        char current_to_wiper( float current ){
+            return char(ceil(float((this->factor*current))));
         }
 
         mbed::I2C* i2c;
-        double currents[8];
+        float currents[8];
 };
